resources/requirements.txt
attrs==22.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.16.2
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.3
jupyterlab-pygments==0.2.2
Markdown==3.4.1
MarkupSafe==2.1.1
mistune==2.0.4
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.2
nest-asyncio==1.5.6
numpy==1.24.1
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
platformdirs==2.6.2
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
six==1.16.0
soupsieve==2.3.2.post1
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.8.1
webencodings==0.5.1
zipp==3.11.0
